chore: hooks moved
