+++ src/bridges/substrate-dvm/config/index.ts
@@ -1,1 +1,2 @@
 export * from './bridge';
+export * from './query';
