+++ src/bridges/substrate-dvm/hooks/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from './records';
