+++ src/bridges/substrate-substrateDVM/hooks/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from './records';
